+++ tests/testthat/test-infer_biocviews.R
@@ -1,25 +1,34 @@
-test_that("infer_biocviews works", {
-
- 
-
+test_that("infer_biocviews works", { 
   # Don't run simply bc biocViews::recommendBiocViews is unable
   ## to find the DESCRIPTION file when running examples.
   # biocviews1 <- infer_biocviews(pkgdir = "../../")
   # testthat::expect_equal(biocviews1,"Software")
+  
+  
+  if(testthat::is_testing()){
+    testthat::expect_error(
+      infer_biocviews(include_branch = FALSE)
+    )  
+  } else {
+    testthat::expect_equal(infer_biocviews(include_branch = FALSE),
+                           "WorkflowManagement")  
+  } 
+  biocviews_manual = c("Software","Genetics","Transcriptomics")
+  biocviews3 <- 
+  if(testthat::is_testing()){
+    testthat::expect_error(
+      infer_biocviews(biocviews = biocviews_manual)
+    )  
+  } else {
+    testthat::expect_equal(infer_biocviews(biocviews = biocviews_manual),
+                           c(biocviews_manual,"WorkflowManagement"))
+  } 
 
-  # biocviews2 <- infer_biocviews(include_branch = FALSE)
-  # testthat::expect_null(biocviews2)
-  # 
-  # biocviews_manual = c("Software","Genetics","Transcriptomics")
-  # biocviews3 <- infer_biocviews(biocviews = biocviews_manual)
-  # testthat::expect_equal(biocviews3,biocviews_manual)
-  # 
-  # #### Errors ####
-  # testthat::expect_error(
-  #   infer_biocviews(branch = "typo")
-  # )
-  # testthat::expect_error(
-  #   infer_biocviews(type = "typo")
-  # )
-  # 
+  #### Errors ####
+  testthat::expect_error(
+    infer_biocviews(branch = "typo")
+  )
+  testthat::expect_error(
+    infer_biocviews(type = "typo")
+  ) 
 })