+++ tests/testthat/test-use_vignette_getstarted.R
@@ -1,8 +1,6 @@
 test_that("use_vignette_getstarted works", {
   
-  if(.Platform$OS.type!="windows"){
-    path <- use_vignette_getstarted(package = "mypackage",
-                                    save_dir = tempdir())
-    testthat::expect_true(file.exists(path))
-  }   
+  path <- use_vignette_getstarted(package = "mypackage",
+                                  save_dir = tempdir())
+  testthat::expect_true(file.exists(path))
 })